requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
yfinance>=0.2.36
vaderSentiment>=3.3.2
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
